Refactor String to use less code space. Add additional compiler warnings and flags typically used in embedded projects

+++ include/EDF/Color.hpp
@@ -8,6 +8,8 @@
 #pragma once
 
 #include "EDF/BitField.hpp"
+#include <cstdint>
+#include <sys/types.h>
 
 namespace EDF {
 
@@ -17,7 +19,7 @@ private:
 public:
     constexpr Color( uint32_t initialValue = 0 ) : rgba( initialValue ) {}
     constexpr Color( uint8_t r, uint8_t g, uint8_t b, uint8_t a = 0xFF ) :
-        rgba( (a << 24) | (r << 16) | (g << 8) | b )
+        rgba( static_cast<uint32_t>((a << 24) | (r << 16) | (g << 8) | b) )
     {}
 
     /* Set color components */
@@ -27,10 +29,10 @@ public:
     constexpr void setA( uint8_t a )          { rgba.set( 24, 8, a ); }
 
     /* Get color components */
-    constexpr uint8_t r()               const { return rgba.get( 16, 8 ); }
-    constexpr uint8_t g()               const { return rgba.get( 8 , 8 ); }
-    constexpr uint8_t b()               const { return rgba.get(  0, 8 ); }
-    constexpr uint8_t a()               const { return rgba.get( 24, 8 ); }
+    constexpr uint8_t r()               const { return static_cast<uint8_t>(rgba.get( 16, 8 )); }
+    constexpr uint8_t g()               const { return static_cast<uint8_t>(rgba.get( 8 , 8 )); }
+    constexpr uint8_t b()               const { return static_cast<uint8_t>(rgba.get(  0, 8 )); }
+    constexpr uint8_t a()               const { return static_cast<uint8_t>(rgba.get( 24, 8 )); }
 
     constexpr uint32_t asRGBA()         const { return rgba.get( 0, 32 ); }
     constexpr uint32_t asRGB()          const { return rgba.get( 0, 24 ); }
@@ -69,10 +71,10 @@ private:
 public:
     constexpr Color565( uint16_t initialValue = 0 ) : rgb( initialValue ) {}
     constexpr Color565( uint8_t r, uint8_t g, uint8_t b ) :
-        rgb( ((r >> 3) << 11) | ((g >> 2) << 5) | (b >> 3) )
+        rgb( static_cast<uint16_t>(((r >> 3) << 11) | ((g >> 2) << 5) | (b >> 3)) )
     {}
     constexpr Color565( const Color& c ) :
-        rgb( ((c.r() >> 3) << 11) | ((c.g() >> 2) << 5) | (c.b() >> 3) )
+        rgb( static_cast<uint16_t>(((c.r() >> 3) << 11) | ((c.g() >> 2) << 5) | (c.b() >> 3)) )
     {}
 
     /* Set color components */
@@ -81,9 +83,9 @@ public:
     constexpr void setB( uint8_t b )          { rgb.set(  0, 5, b >> 3 ); }
 
     /* Get color components */
-    constexpr uint8_t r()               const { return rgb.get( 11, 5 ) << 3; }
-    constexpr uint8_t g()               const { return rgb.get(  5, 6 ) << 2; }
-    constexpr uint8_t b()               const { return rgb.get(  0, 5 ) << 3; }
+    constexpr uint8_t r()               const { return static_cast<uint8_t>(rgb.get( 11, 5 ) << 3); }
+    constexpr uint8_t g()               const { return static_cast<uint8_t>(rgb.get(  5, 6 ) << 2); }
+    constexpr uint8_t b()               const { return static_cast<uint8_t>(rgb.get(  0, 5 ) << 3); }
 
     constexpr uint16_t asRGB()          const { return rgb.get( 0, 16 ); }
 };
